requirements.txt
Flask
requests
numpy
pillow
flask-caching
flask-injector
paho-mqtt==1.5.1
flask-socketio
blinker